refactor: domain model
