+++ internal/infrastructure/repository/postRepository.go
@@ -34,6 +34,64 @@ func toPost(gormPost *model.Post) *postDomain.Post {
 	}
 }
 
+func toGormComment(comment *postDomain.Comment) *model.Comment {
+	return &model.Comment{
+		ID:        comment.CommentId.String(),
+		PostID:    comment.PostId.String(),
+		UserID:    comment.Commenter.UserId.String(),
+		Content:   comment.Content.String(),
+		CreatedAt: comment.CreatedAt,
+	}
+}
+
+func toComment(gormComment *model.Comment) *postDomain.Comment {
+	// TODO: fix N+1 problem
+	commenter, _ := userDomain.Factory.GetUser(userDomain.UserId(gormComment.UserID))
+
+	// TODO: fix N+1 problem
+	gormReplies, _ := query.Reply.WithContext(context.Background()).Where(query.Reply.CommentID.Eq(gormComment.ID)).Find()
+
+	var replies []*postDomain.Reply
+	for _, gormReply := range gormReplies {
+		// TODO: fix N+1 problem
+		replies = append(replies, toReply(gormReply))
+	}
+
+	return &postDomain.Comment{
+		CommentId: postDomain.CommentId(gormComment.ID),
+		PostId:    postDomain.PostId(gormComment.PostID),
+		Commenter: commenter,
+		Content:   postDomain.Content(gormComment.Content),
+		Replies:   replies,
+		CreatedAt: gormComment.CreatedAt,
+	}
+}
+
+func toGormReply(reply *postDomain.Reply) *model.Reply {
+	return &model.Reply{
+		ID:        reply.ReplyId.String(),
+		CommentID: reply.CommentId.String(),
+		UserID:    reply.Replier.UserId.String(),
+		Sequence:  reply.Sequence.Int32(),
+		Content:   reply.Content.String(),
+		CreatedAt: reply.CreatedAt,
+	}
+}
+
+func toReply(gormReply *model.Reply) *postDomain.Reply {
+	// TODO: fix N+1 problem
+	replier, _ := userDomain.Factory.GetUser(userDomain.UserId(gormReply.UserID))
+
+	return &postDomain.Reply{
+		ReplyId:   postDomain.ReplyId(gormReply.ID),
+		CommentId: postDomain.CommentId(gormReply.CommentID),
+		Replier:   replier,
+		Sequence:  postDomain.Sequence(gormReply.Sequence),
+		Content:   postDomain.Content(gormReply.Content),
+		CreatedAt: gormReply.CreatedAt,
+	}
+}
+
 func (r *PostRepository) Create(post *postDomain.Post) error {
 	gormPost := toGormPost(post)
 	gormPost.ID = xid.New().String()
@@ -46,7 +104,7 @@ func (r *PostRepository) Delete(post *postDomain.Post) error {
 	return err
 }
 
-func (r *PostRepository) FindById(postId postDomain.PostId) (*postDomain.Post, error) {
+func (r *PostRepository) FindPostById(postId postDomain.PostId) (*postDomain.Post, error) {
 	if !r.IsExistPostId(postId) {
 		return nil, errors.New("post not found")
 	}
@@ -97,4 +155,49 @@ func (r *PostRepository) GetLikers(post *postDomain.Post) ([]*userDomain.User, e
 	}
 
 	return userDomain.Factory.GetUsers(likerUserIdList)
+}
+
+func (r *PostRepository) CreateComment(comment *postDomain.Comment) error {
+	gormComment := toGormComment(comment)
+	return query.Comment.WithContext(context.Background()).Save(gormComment)
+}
+
+func (r *PostRepository) DeleteComment(comment *postDomain.Comment) error {
+	_, err := query.Comment.WithContext(context.Background()).Where(query.Comment.ID.Eq(comment.CommentId.String())).Delete()
+	return err
+}
+
+func (r *PostRepository) CreateReply(reply *postDomain.Reply) error {
+	gormReply := toGormReply(reply)
+	return query.Reply.WithContext(context.Background()).Save(gormReply)
+}
+
+func (r *PostRepository) DeleteReply(reply *postDomain.Reply) error {
+	_, err := query.Reply.WithContext(context.Background()).Where(query.Reply.ID.Eq(reply.ReplyId.String())).Delete()
+	return err
+}
+
+func (r *PostRepository) FindCommentByPostId(postId postDomain.PostId) ([]*postDomain.Comment, error) {
+	gormComments, err := query.Comment.WithContext(context.Background()).Where(query.Comment.PostID.Eq(postId.String())).Find()
+	if err != nil {
+		return nil, err
+	}
+
+	var comments []*postDomain.Comment
+	for _, gormComment := range gormComments {
+		// TODO: fix N+1 problem
+		comments = append(comments, toComment(gormComment))
+	}
+
+	return comments, nil
+}
+
+func (r *PostRepository) IsExistCommentId(commentId postDomain.CommentId) bool {
+	count, _ := query.Comment.WithContext(context.Background()).Where(query.Comment.ID.Eq(commentId.String())).Count()
+	return count != 0
+}
+
+func (r *PostRepository) IsExistReplyId(replyId postDomain.ReplyId) bool {
+	count, _ := query.Reply.WithContext(context.Background()).Where(query.Reply.ID.Eq(replyId.String())).Count()
+	return count != 0
 }
